Initial support for per-colorscheme colors
I was initially distracted by this plugin because (I think) of the
contrast between the colors I'm used to seeing and those in the
colorscheme I currently use (solarized dark).

To fix this, I just check if we have a set of colors that matches the
users' current colorscheme (something like s:<colorscheme>_colours or
s:<colorscheme>_<light|dark>_colours) and use that if the user hasn't
already explicitly requested the legacy colors.

I've added initial support for solarized.  I'm not using all the colors
because some of them had too much contrast (the white ones) or to little
(the black ones).

It'd be nice to have support for some popular themes (vividchalk,
monokai, github, etc), and possibly built-in themes (these might need to
be tested with 8 or 16 colors).